move BenefitsForm to profile
